# Flsurf/Client/FlsurfSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SpakOfMind Flsurf",
    "version": "v1"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserSchema"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserSchema"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserSchema"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserSchema"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/auth/external-login/{provider}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "ExternalLogin",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginUserSchema": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
...
      "RegisterUserSchema": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "country",
          "email",
          "name",
          "password",
          "surname",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 52,
            "minLength": 0
          },
          "surname": {
            "type": "string",
            "maxLength": 52,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "country": {
            "type": "string",
            "enum": [
              "Kazakhstan",
              "Russia",
              "Belarus"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "Freelancer",
              "Client",
              "Other",
              "NonUser",
              "Staff"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Session Token",
        "name": "session_token",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}